ok
ok

+++ Started/Models/Employees.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -7,10 +8,10 @@ namespace Started.Models
 {
     public class Employees
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string Name { get; set; }
         public string Surname { get; set; }
         public string Job { get; set; }
-        public int Salary { get; set; }
+        public double Salary { get; set; }
     }
 }
